Fix Routing and Update Styles
Fixed routing issues by adhering to nextjs standard.
Updated button styles to make UI look better
