+++ oml-gpt/src/app/Components/ButtonWithLink.jsx
@@ -1,0 +1,9 @@
+import Link from 'next/link'
+
+export default function ButtonWithLink({ href, children }) {
+    return (
+        <Link href={ href }>
+            <button class="btn btn-outline btn-accent">{ children }</button>
+        </Link>
+    )
+}
